+++ XMonad/Util/PureX.hs
@@ -44,6 +44,7 @@ module XMonad.Util.PureX (
   withWindowSet', withFocii,
   modify'', modifyWindowSet',
   getStack, putStack, peek,
+  focusWindow,
   view, greedyView, invisiView,
   shift, curScreen, curWorkspace,
   curTag, curScreenId,
@@ -271,6 +272,18 @@ shift tag = withFocii $ \ctag fw ->
     modifyWindowSet' (W.shiftWin tag fw)
     mfw' <- peek
     return (Any $ Just fw /= mfw')
+
+-- | Internal. Refresh-tracking logic of focus operations.
+focusWith :: XLike m => (WindowSet -> WindowSet) -> m Any
+focusWith focuser = do
+    old <- peek
+    modifyWindowSet' focuser
+    new <- peek
+    return (Any $ old /= new)
+
+-- | A refresh-tracking version of @W.focusWindow@.
+focusWindow :: XLike m => Window -> m Any
+focusWindow w = focusWith (W.focusWindow w)
 
 -- }}}